requirements.txt
Flask
Flask-Bootstrap4
Redis
